fix undeterministic behavior: we were reading *cur while (cur == end) could be true
This bug appeared after adding code that in all cases makes FSST (re-)consider single-byte symbols and single-byte extensions of codes. Lot's of things were tried, this was one of them. It gives FSST a way of quickly falling back to shorter codes, if longer codes lose out in the gain competetition. But in the haste of many experiments and many things that were tried, this error slipped in.

The code, in general, but certainly here, could also be more clean. I will follow-up with a small cleanup.