Make `fsst_compress` and `fsst_decompress` const-correct
Previously, the input parameters to `fsst_compress` and `fsst_compress` were passed in as `size_t[]` and `unsigned char*[]` respectively although these parameters are never modified. This commit changes these types to `const size_t[]` and `const unsigned char*[]`. For the same reason, the decoder is now passed as a `const fsst_decoder_t*` to the `fsst_decompress` function (it previously was a `fsst_deocder_t*`.

This commit also removes three unused variables, that lead to compiler warnings.